Removed all old freelex code
