--- app/models/freelex/example.rb
@@ -1,6 +1,0 @@
-module Freelex
-  class Example
-    include ActiveModel::Model
-    attr_accessor :video, :translation, :sentence
-  end
-end
