+++ app/views/admin/requests/index.html.haml
@@ -1,4 +1,4 @@
-%h1 Requests to Freelex
+%h1 Requests to Signbank
 = will_paginate(@requests)
 
 %table
