+++ app/views/layouts/admin.html.haml
@@ -64,7 +64,7 @@
                 = link_to admin_requests_path do
                   %span.admin-dashboard-sidebar-text
                     %i.fi-list-bullet
-                    Freelex Logs
+                    Signbank Logs
               %li
                 = link_to edit_admin_user_path do
                   %span.admin-dashboard-sidebar-text
