+++ db/seeds/settings.csv
@@ -2,4 +2,3 @@
 6,glossary,10,2011-09-26 00:56:54.839789,2011-10-26 21:09:25.586241
 7,404,12,2011-09-26 00:56:54.856156,2011-09-26 00:56:54.863798
 8,footer,11,2011-09-26 00:56:54.872765,2011-09-26 00:56:54.880427
-10,dictionary_adapter,freelex,2022-11-28 00:00:00,2022-11-28 00:00:00
