+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "nzsl-online",
   "version": "1.0.0",
-  "description": "New Zealand Sign Language Dictionary consists of 2 major units * Ruby on Rails Website (this repo) * Freelex, and open source project for maintaining the signs lexicon data.",
+  "description": "The Ruby on Rails website for the New Zealand Sign Language Dictionary.",
   "main": "index.js",
   "directories": {
     "doc": "doc"
