+++ spec/controllers/signbank/signs_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe Signbank::SignsController, type: :controller, sign_model_adapter: :signbank do
+RSpec.describe Signbank::SignsController, type: :controller do
   describe '/' do
     before { Rails.application.load_seed }
     render_views
